Implemented /sales and /price endpoints, req. DTO, and changes to related files.

+++ src/main/java/com/rbanerjee/SmartSabzi/Entity/SaleItem.java
@@ -26,7 +26,7 @@ public class SaleItem {
     private VendorVegetable vendorVegetable;
 
     @Column
-    private double weight;
+    private BigDecimal weight;
 
     @Column
     private BigDecimal unitPrice;
